pyproject.toml
[tool.poetry]
name = "2022-platform"
version = "0.1.0"
description = ""
authors = ["Boris Rakovan <boris.rakovan@ataccama.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.85.1"
uvicorn = "^0.19.0"
alembic = "^1.8.1"
sqlmodel = "^0.0.8"
tortilla = "^0.5.0"
python-dotenv = "^0.21.0"
numpy = "^1.23.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
